fix a warning in stressMathore and a problem making dictionary library on Windows
git-svn-id: http://root.cern.ch/svn/root/trunk@21313 27541ba8-7e3a-0410-8455-c3a389f83636

+++ test/stressMathMore.cxx
@@ -542,7 +542,7 @@ int stressMathMore(double nscale = 1) {
    TBenchmark bm;
    bm.Start("stressMathMore");
    
-   int n = nscale*100;
+   int n = int(nscale*100);
 
    iret |= testGammaFunction(n);
    iret |= testBetaFunction(n);
